Feat/friends

+++ db/migrations/000003_create_friends_table.up.sql
@@ -1,7 +1,7 @@
 create table if not exists friends(
-    id serial primary key,
-    user_id int not null references users(id),
-    friend_id int not null references users(id),
+    id BIGSERIAL primary key,
+    user_id BIGSERIAL not null references users(id),
+    friend_id BIGSERIAL not null references users(id),
     created_at timestamptz not null default current_timestamp
 );
